refactor(download): move the grab pipeline and its routes into plugins/download
The movie and episode download services, the torrent history and auto matchers, the download-history
entity and the eight grab routes now live under `plugins/download/`. The routes keep their exact
method and path — the frontend and every frozen client call them, so a new controller declares the
same prefix beside the one that keeps core's other routes.

`auto-grab-pipeline` splits where the plan draws the line: `buildScoringContext`,
`classifyForSearch` and `shouldSearchMissing` stay in core because they read core table columns and
an eagerly-loaded quality profile, while the executor moves. The scheduler now injects both, and the
scoring kernel in `common/release-scoring/` stays put — it is shared with profiles and media-query,
and its rejection codes are a frontend ABI.

Breaks the import cycle that made this boot fail: the supervisor took `CURRENT_FLIKS_VERSION` from
the registry service, which the registry reaches through the process service, so once module
evaluation order shifted the registry saw an undefined dependency. It now reads the leaf module that
exists for exactly this. The indexers module also dropped an import of the plugins module that
nothing in it used and that closed a second cycle through the policy vocabulary.

The fence allowlist is re-synced: the files that moved in are out of it, and the one core file that
newly reaches across is in.

Refs: #894
