+++ backend/src/modules/blocklist/blocklist.service.ts
@@ -6,7 +6,7 @@ import { CreateBlocklistEntryDto } from './dto/create-blocklist-entry.dto';
 import { Indexer } from '../../plugins/download/indexers/entities/indexer.entity';
 import { Media } from '../media/entities/media.entity';
 import { User } from '../users/entities/user.entity';
-import { DownloadHistory } from '../media/entities/download-history.entity';
+import { DownloadHistory } from '../../plugins/download/entities/download-history.entity';
 
 @Injectable()
 export class BlocklistService {
