+++ backend/src/modules/counts/counts.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { DownloadHistory } from '../media/entities/download-history.entity';
+import { DownloadHistory } from '../../plugins/download/entities/download-history.entity';
 import { FliksRequest } from '../requests/entities/request.entity';
 import { CountsService } from './counts.service';
 import { CountsController } from './counts.controller';
