+++ backend/src/modules/scheduler/completion.service.spec.ts
@@ -5,11 +5,11 @@ import { DataSource } from 'typeorm';
 import { CompletionService } from './completion.service';
 import { AcquisitionEventsService } from './acquisition-events.service';
 import { NamingService } from './naming.service';
-import { DownloadHistory } from '../media/entities/download-history.entity';
+import { DownloadHistory } from '../../plugins/download/entities/download-history.entity';
 import { Media } from '../media/entities/media.entity';
 import { Library } from '../libraries/entities/library.entity';
 import { QbittorrentTorrent } from '../../plugins/download/download-clients/qbittorrent.service';
-import { TorrentHistoryMatcher } from '../media/torrent-history-matcher.service';
+import { TorrentHistoryMatcher } from '../../plugins/download/torrent-history-matcher.service';
 import { FileTransferService } from '../../common/services/file-transfer.service';
 import { LibraryIngestService } from '../../common/library-ingest/library-ingest.service';
 import { MediaType } from '../../common/enums';
