+++ backend/src/plugins/download/download-clients/download-clients.service.ts
@@ -2,7 +2,7 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { In, Repository } from 'typeorm';
 import { DownloadClient } from './entities/download-client.entity';
-import { DownloadHistory } from '../../../modules/media/entities/download-history.entity';
+import { DownloadHistory } from '../entities/download-history.entity';
 import { Media } from '../../../modules/media/entities/media.entity';
 import { StalledCheck } from '../../../modules/scheduler/entities/stalled-check.entity';
 import { CleanupProfile } from '../../../modules/cleanup-profiles/entities/cleanup-profile.entity';
@@ -15,7 +15,7 @@ import { QbittorrentService, QbittorrentTorrent } from './qbittorrent.service';
 import { CreateDownloadClientDto } from './dto/create-download-client.dto';
 import { UpdateDownloadClientDto } from './dto/update-download-client.dto';
 import { TestDownloadClientDto } from './dto/test-download-client.dto';
-import { TorrentHistoryMatcher } from '../../../modules/media/torrent-history-matcher.service';
+import { TorrentHistoryMatcher } from '../torrent-history-matcher.service';
 import { BlocklistService } from '../../../modules/blocklist/blocklist.service';
 import { EventsService } from '../../../modules/scheduler/events.service';
 import { redactSecretFields, mergeSecretFields } from '../../../common/utils/secret-fields.util';
